# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  financetracker_internal_application_dto.CategoryListResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/financetracker_internal_application_dto.CategoryResponse'
        type: array
      total_count:
        type: integer
    type: object
  financetracker_internal_application_dto.CategoryMasterListResponse:
    properties:
      category_masters:
        items:
          $ref: '#/definitions/financetracker_internal_application_dto.CategoryMasterResponse'
        type: array
      total_count:
        type: integer
    type: object
  financetracker_internal_application_dto.CategoryMasterResponse:
    properties:
      category_type:
        type: string
      color:
        type: string
      created_at:
        type: string
      display_order:
        type: integer
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  financetracker_internal_application_dto.CategoryResponse:
    properties:
      category_master_id:
        type: string
      created_at:
        type: string
      custom_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  financetracker_internal_application_dto.CreateCategoryRequest:
    properties:
      category_master_id:
        type: string
      custom_name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - category_master_id
    type: object
...
  financetracker_internal_application_dto.UpdateCategoryRequest:
    properties:
      custom_name:
        maxLength: 50
        minLength: 1
        type: string
      is_active:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@financetracker.local
    name: API Support
  description: Finance management REST API with Auth0 authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FinanceTracker API
  version: "1.0"
paths:
...
  /api/v1/categories:
    get:
      consumes:
      - application/json
      description: ユーザーのカスタムカテゴリ一覧を取得します
      parameters:
      - description: カテゴリマスターID
        in: query
        name: category_master_id
        type: string
      - description: アクティブ状態
        in: query
        name: is_active
        type: boolean
      - default: '"created_at desc"'
        description: ソート順
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.CategoryListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: ユーザーカテゴリ一覧取得
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: 新しいユーザーカテゴリを作成します
      parameters:
      - description: 作成するカテゴリ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/financetracker_internal_application_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: カテゴリ作成
      tags:
      - categories
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: 指定されたIDのカテゴリを削除（非活性化）します
      parameters:
      - description: カテゴリID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: カテゴリ削除
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: 指定されたIDのカテゴリを取得します
      parameters:
      - description: カテゴリID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: カテゴリ取得
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: 指定されたIDのカテゴリを更新します
      parameters:
      - description: カテゴリID
        in: path
        name: id
        required: true
        type: string
      - description: 更新するカテゴリ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/financetracker_internal_application_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: カテゴリ更新
      tags:
      - categories
  /api/v1/categories/master:
    get:
      consumes:
      - application/json
      description: 利用可能なカテゴリマスター（テンプレート）一覧を取得します
      parameters:
      - description: カテゴリタイプ
        enum:
        - income
        - expense
        in: query
        name: category_type
        type: string
      - default: '"display_order asc"'
        description: ソート順
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/financetracker_internal_application_dto.CategoryMasterListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: カテゴリマスター一覧取得
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"